Move inputs and outputs to separate file
