+++ Localize/lang/strings.json
@@ -20,6 +20,7 @@
   "/VcZ9g": "This connector has multiple versions, built-in and Azure-hosted. Use built-in for the best performance, connection-string authentication, and other features. Use Azure-hosted for other authentication options.",
   "/WW7If": "True",
   "/doURb": "Convert the input to an array",
+  "/mjH84": "Show raw outputs",
   "/n13VL": "Properties",
   "/qchXQ": "Required. The collection to join items from.",
   "/qu3zt": "0",
@@ -535,7 +536,6 @@
   "U3guPp": "Group by Connector",
   "U3iWVd": "Generates an array of integers starting from a certain number",
   "U56j9a": "(UTC-06:00) Easter Island",
-  "UAwAAS": "Show inputs",
   "UCNM4L": "To reference a parameter, use the dynamic content list.",
   "UEryJE": "Delete Workflow Action",
   "UHCVNK": "Replaces a string with a given string",
@@ -566,7 +566,6 @@
   "VptXzY": "Use \"{value}\" as a custom value",
   "W070M2": "of {max}",
   "W6FdMh": "Required. The name of the new property.",
-  "WCrlES": "Show outputs",
   "WGwH45": "Clear",
   "WLj8DH": "Please enter a valid dictionary",
   "WOOYW1": "Actions perform operations on data or communicate between systems to execute.",
@@ -647,6 +646,7 @@
   "_/VcZ9g.comment": "Description about hybrid connectors",
   "_/WW7If.comment": "True",
   "_/doURb.comment": "Label for description of custom array Function",
+  "_/mjH84.comment": "Show outputs text",
   "_/n13VL.comment": "Properties text",
   "_/qchXQ.comment": "Required collection parameter to apply join function on",
   "_/qu3zt.comment": "Hour of the day",
@@ -1162,7 +1162,6 @@
   "_U3guPp.comment": "Label for the checkbox to group results by connector",
   "_U3iWVd.comment": "Label for description of custom range Function",
   "_U56j9a.comment": "Timezone value ",
-  "_UAwAAS.comment": "Show inputs text",
   "_UCNM4L.comment": "Description for Workflow Parameters Part 2",
   "_UEryJE.comment": "Title for operation node",
   "_UHCVNK.comment": "Label for description of custom replace Function",
@@ -1193,7 +1192,6 @@
   "_VptXzY.comment": "Label for button to allow user to create custom value in combobox from current input",
   "_W070M2.comment": "Text on a pager where people can select a page number out of {max}",
   "_W6FdMh.comment": "Required string parameter for new property name in addProperty function",
-  "_WCrlES.comment": "Show outputs text",
   "_WGwH45.comment": "Label to clear editor",
   "_WLj8DH.comment": "Error Message for Invalid Dictionary",
   "_WOOYW1.comment": "Description of what Actions are, on a tooltip about Actions",
@@ -1646,6 +1644,7 @@
   "_xN3GEX.comment": "Client Certificate Password Placeholder Text",
   "_xOAOHM.comment": "Label for input Field for String Editor",
   "_xQQ9ko.comment": "title for pagination user input",
+  "_xSMbKr.comment": "Show inputs text",
   "_xV/oyK.comment": "Required parameter for new/existing property value in setProperty function",
   "_xV4Koe.comment": "Label for code tab",
   "_xYyPR8.comment": "Label for description of custom uriPath Function",
@@ -2076,6 +2075,7 @@
   "xN3GEX": "Enter password as plain text or use a secure parameter",
   "xOAOHM": "Editor Input",
   "xQQ9ko": "Threshold",
+  "xSMbKr": "Show raw inputs",
   "xV/oyK": "Required. The name of the new or existing property.",
   "xV4Koe": "Code",
   "xYyPR8": "Returns the path from a URI. If path is not specified, returns '/'",
