+++ libs/designer/src/lib/ui/panel/panelTabs/monitoringTab/monitoringTab.tsx
@@ -2,103 +2,23 @@ import constants from '../../../../common/constants';
 import { useSelectedNodeId } from '../../../../core/state/panel/panelSelectors';
 import { useBrandColor } from '../../../../core/state/selectors/actionMetadataSelector';
 import { useNodeMetadata } from '../../../../core/state/workflow/workflowSelectors';
+import { InputsPanel } from './inputsPanel';
+import { OutputsPanel } from './outputsPanel';
 import { PropertiesPanel } from './propertiesPanel';
-import { ErrorSection, ValuesPanel } from '@microsoft/designer-ui';
+import { ErrorSection } from '@microsoft/designer-ui';
 import type { PanelTab } from '@microsoft/designer-ui';
-import { useIntl } from 'react-intl';
 
 export const MonitoringPanel: React.FC = () => {
-  const intl = useIntl();
   const selectedNodeId = useSelectedNodeId();
   const nodeMetadata = useNodeMetadata(selectedNodeId);
   const brandColor = useBrandColor(selectedNodeId);
   const runMetaData = nodeMetadata?.runData;
 
-  const intlText = {
-    inputs: intl.formatMessage({
-      defaultMessage: 'Inputs',
-      description: 'Inputs text',
-    }),
-    noInputs: intl.formatMessage({
-      defaultMessage: 'No inputs',
-      description: 'No inputs text',
-    }),
-    showInputs: intl.formatMessage({
-      defaultMessage: 'Show inputs',
-      description: 'Show inputs text',
-    }),
-    outputs: intl.formatMessage({
-      defaultMessage: 'Outputs',
-      description: 'Outputs text',
-    }),
-    noOutputs: intl.formatMessage({
-      defaultMessage: 'No outputs',
-      description: 'No utputs text',
-    }),
-    showOutputs: intl.formatMessage({
-      defaultMessage: 'Show outputs',
-      description: 'Show outputs text',
-    }),
-  };
-
   return (
     <div>
       <ErrorSection className="msla-request-history-panel-error" error={runMetaData?.error} />
-      {runMetaData?.inputsLink ? (
-        <ValuesPanel
-          brandColor={brandColor}
-          headerText={intlText.inputs}
-          linkText={intlText.showInputs}
-          showLink={true}
-          values={{
-            method: {
-              displayName: 'Method',
-              value: 'POST',
-            },
-            uri: {
-              displayName: 'URL',
-              value: 'https://httpbin.org/post/',
-            },
-          }}
-          labelledBy={''}
-          noValuesText={intlText.noInputs}
-          showMore={false}
-        />
-      ) : null}
-      {runMetaData?.outputsLink ? (
-        <ValuesPanel
-          brandColor={brandColor}
-          headerText={intlText.outputs}
-          linkText={intlText.showOutputs}
-          showLink={true}
-          values={{
-            statusCode: {
-              displayName: 'Status code',
-              value: 200,
-            },
-            headers: {
-              displayName: 'Headers',
-              format: 'key-value-pairs',
-              value: {
-                Date: 'Fri, 28 Jan 2022 00:02:51 GMT',
-                Expires: '-1',
-                Pragma: 'no-cache',
-                Vary: 'Accept-Encoding',
-              },
-            },
-            body: {
-              displayName: 'Body',
-              value: {
-                nextLink: '[REDACTED]',
-                value: [],
-              },
-            },
-          }}
-          labelledBy={''}
-          noValuesText={intlText.noOutputs}
-          showMore={false}
-        />
-      ) : null}
+      <InputsPanel runMetaData={runMetaData} brandColor={brandColor} />
+      <OutputsPanel runMetaData={runMetaData} brandColor={brandColor} />
       <PropertiesPanel properties={runMetaData} brandColor={brandColor} />
     </div>
   );
