+++ libs/services/designer-client-services/src/lib/host.ts
@@ -1,0 +1,11 @@
+export enum ContentType {
+  Inputs,
+  Outputs,
+}
+
+/**
+ * The host service.
+ */
+export interface IHostService {
+  fetchAndDisplayContent?(identifier: string, title: string, url: string, type: ContentType): Promise<void>;
+}
