+++ libs/services/intl/src/compiled-lang/strings.en-XA.json
@@ -125,6 +125,12 @@
       "value": "Ḉŏńṿèŕṭ ṭḫè íńṗůṭ ṭŏ âń âŕŕâẏ"
     }
   ],
+  "/mjH84": [
+    {
+      "type": 0,
+      "value": "Ṡḫŏẘ ŕâẘ ŏůṭṗůṭś"
+    }
+  ],
   "/n13VL": [
     {
       "type": 0,
@@ -3509,12 +3515,6 @@
       "value": "(ŨṮḈ-06:00) Ḕâśṭèŕ Ḭśĺâńḋ"
     }
   ],
-  "UAwAAS": [
-    {
-      "type": 0,
-      "value": "Ṡḫŏẘ íńṗůṭś"
-    }
-  ],
   "UCNM4L": [
     {
       "type": 0,
@@ -3709,12 +3709,6 @@
     {
       "type": 0,
       "value": "Ŕèɋůíŕèḋ. Ṯḫè ńâṁè ŏḟ ṭḫè ńèẘ ṗŕŏṗèŕṭẏ."
-    }
-  ],
-  "WCrlES": [
-    {
-      "type": 0,
-      "value": "Ṡḫŏẘ ŏůṭṗůṭś"
     }
   ],
   "WGwH45": [
@@ -6663,6 +6657,12 @@
     {
       "type": 0,
       "value": "Ṯḫŕèśḫŏĺḋ"
+    }
+  ],
+  "xSMbKr": [
+    {
+      "type": 0,
+      "value": "Ṡḫŏẘ ŕâẘ íńṗůṭś"
     }
   ],
   "xV/oyK": [
