+++ libs/services/intl/src/compiled-lang/strings.json
@@ -125,6 +125,12 @@
       "value": "Convert the input to an array"
     }
   ],
+  "/mjH84": [
+    {
+      "type": 0,
+      "value": "Show raw outputs"
+    }
+  ],
   "/n13VL": [
     {
       "type": 0,
@@ -3509,12 +3515,6 @@
       "value": "(UTC-06:00) Easter Island"
     }
   ],
-  "UAwAAS": [
-    {
-      "type": 0,
-      "value": "Show inputs"
-    }
-  ],
   "UCNM4L": [
     {
       "type": 0,
@@ -3709,12 +3709,6 @@
     {
       "type": 0,
       "value": "Required. The name of the new property."
-    }
-  ],
-  "WCrlES": [
-    {
-      "type": 0,
-      "value": "Show outputs"
     }
   ],
   "WGwH45": [
@@ -6663,6 +6657,12 @@
     {
       "type": 0,
       "value": "Threshold"
+    }
+  ],
+  "xSMbKr": [
+    {
+      "type": 0,
+      "value": "Show raw inputs"
     }
   ],
   "xV/oyK": [
